requirements.txt
google-generativeai
python-dotenv
speechrecognition
pydub
simpleaudio
gTTS
PyAudio
pycaw
screen-brightness-control
pyttsx3